Add multi-ps in client + use shared memory for data in server

+++ PowerSupply/server.c
@@ -18,11 +18,14 @@
 #include <sys/socket.h>
 #include <sys/types.h>
 #include <sys/wait.h>
+#include <sys/shm.h>
+#include <sys/ipc.h>
 
 #define POWER_THRESHOLD 5000
 #define WARNING_THRESHOLD 4500
 #define BACKLOG 10 /* Number of allowed connections */
 #define BUFF_SIZE 8192
+#define MAX_DEVICE 10
 
 ////////////////////
 // Variables list //
@@ -36,6 +39,19 @@ int bytes_sent, bytes_received;
 struct sockaddr_in server;
 struct sockaddr_in client;
 int sin_size;
+char use_mode[][10] = {"off", "normal", "limited"};
+key_t key = 1234;
+int shmid;
+
+typedef struct {
+	int pid;
+	char name[50];
+	int normal;
+	int limit;
+	int mode;
+} device_t;
+
+device_t *devices;
 
 int tprintf(const char* fmt, ...){
 	va_list args;
@@ -49,15 +65,31 @@ int tprintf(const char* fmt, ...){
 
 void powerSupply_handle(int conn_sock) {
 	int is_first_message = 1;
-	char name[50];
-	int mode_2;
-	int mode_3;
-	char use_mode[3][10] = {"off", "normal", "limited"};
+	int no;
 
+	// get memory to store device infomation
+	if ((devices = (device_t*) shmat(shmid, (void*) 0, 0)) == (void*)-1) {
+		tprintf("shmat() failed\n");
+		exit(1);
+	}
+	for (no = 0; no < MAX_DEVICE; no++) {
+		if (devices[no].pid == 0)
+			break;
+	}
 	while(1) {
+
+		///////////////////
+		// listen on tcp //
+		///////////////////
 		bytes_received = recv(conn_sock, recv_data, BUFF_SIZE-1, 0);
 		if (bytes_received <= 0) {
-			tprintf("Device [%s] disconnected\n\n", name);
+			// if DISCONNECT
+			tprintf("Device [%s] disconnected\n\n", devices[no].name);
+			devices[no].pid = 0;
+			strcpy(devices[no].name, "");
+			devices[no].normal = 0;
+			devices[no].limit = 0;
+			devices[no].mode = 0;
 			powerSupply_count--;
 			kill(getpid(), SIGKILL);
 			break;
@@ -65,27 +97,26 @@ void powerSupply_handle(int conn_sock) {
 			recv_data[bytes_received] = '\0';
 		}
 		
-		////////////////////////////
-		// Create device instance //
-		////////////////////////////
-		
+		// Create device instance for first time receive
 		if (is_first_message) {
 			is_first_message = 0;
-			sscanf(recv_data, "%[^|]|%d|%d", name, &mode_2, &mode_3);
+			sscanf(recv_data, "%[^|]|%d|%d", devices[no].name, &devices[no].normal, &devices[no].limit);
+			devices[no].pid = getpid();
+			devices[no].mode = 0;
 			tprintf("--- Connected device info ---\n");
-			tprintf("       name: %s\n", name);
-			tprintf("     normal: %dW\n", mode_2);
-			tprintf("      limit: %dW\n", mode_3);
-			tprintf("   use mode: %s\n", use_mode[0]);
+			tprintf("       name: %s\n", devices[no].name);
+			tprintf("     normal: %dW\n", devices[no].normal);
+			tprintf("      limit: %dW\n", devices[no].limit);
+			tprintf("   use mode: %s\n", use_mode[devices[no].mode]);
 			tprintf("-----------------------------\n\n");
 			continue;
 		}
 
-		int menu = atoi(recv_data);
-		tprintf("%d\n",menu);
+		devices[no].mode = atoi(recv_data);
+		tprintf("Device [%s] change mode to [%s]\n", devices[no].name, use_mode[devices[no].mode]);
 
 	} // endwhile
-}
+} // end function
 
 void connectMng_handle() {
 
@@ -139,7 +170,7 @@ void connectMng_handle() {
 			//parent
 			close(conn_sock);
 			powerSupply_count++;
-			tprintf("A device connected, connectMng forked process powerSupply(%d) --- pid: %d.\n", powerSupply_count, powerSupply);
+			tprintf("A device connected, connectMng forked process powerSupply --- pid: %d.\n", powerSupply);
 		}
 	} //end communication
 
@@ -147,9 +178,15 @@ void connectMng_handle() {
 } //end function
 
 void elePowerCtrl_handle() {
+	while(1) {
+
+	} // endwhile
 }
 
 void powSupplyInfoAccess_handle() {
+	while(1) {
+		
+	} // endwhile
 }
 
 void logWrite_handle() {
@@ -163,6 +200,15 @@ int main(int argc, char const *argv[])
 	}
 	server_port = atoi(argv[1]);
 	printf("SERVER start, PID is %d.\n", getpid());
+
+	///////////////////////////////////
+	// Create shared memory for data //
+	///////////////////////////////////
+
+	if ((shmid = shmget(key, sizeof(device_t) * MAX_DEVICE, 0644 | IPC_CREAT)) < 0) {
+		tprintf("shmge() failed\n");
+		exit(1);
+	}
 
 	///////////////////////////////////
 	// start child process in SERVER //
